refactor(medication): Reorder `remark` field in `MedicationRecord` and `MedicationRecordEntity`
This commit changes the order of the `remark` field in the `MedicationRecord` data class and its corresponding `MedicationRecordEntity` mapper.

In `MedicationRecord.kt`, the `remark` field has been moved after `timeZone`.
In `MedicationRecordMapper.kt`, the `remark` field in both `toDomain()` and `toEntity()` functions has been moved to align with the change in the `MedicationRecord` data class.

Signed-off-by: 339 <339@339.im>