feat(medication): Make medication stock optional
The `stock` property in the `Medication` domain model and `MedicationEntity` has been changed from `Float` to `Float?`, allowing it to be nullable.

This change is reflected in the database schema, where the `stock` column in the `medication` table is no longer `NOT NULL`.

In `AddMedicationViewModel`, the handling of stock input has been updated. If the stock string cannot be parsed to a float, it will now be set to `null` instead of a default value. The `DEFAULT_STOCK` constant has been removed.

Signed-off-by: 339 <339@339.im>