refactor: Allow selection of past dates in DateTextField
This commit modifies the `DateTextField` to allow the selection of dates in the past by default.

Previously, the `DateTextField` restricted selections to today and future dates by default (`FromTodaySelectableDates`). This has been changed to `DatePickerDefaults.AllDates`, making the component more flexible for various use cases.

Additionally, in the `ScheduleTypeStep` of the medication plan creation flow, the selectable range for the end date is now correctly calculated to start from the later of either the selected start date or today's date. The `currentDate` in the `DatePickerDialog` is also now nullable to align with the state.

Signed-off-by: 339 <339@339.im>