feat(medication-plan): Add instructional text to empty medication plan screen
This commit enhances the user experience on the medication plan screen when it is empty.

A new instructional text, "Click the '+' button to create your first medication plan," has been added below the "No medication plans yet" message. This provides clear guidance to new users on how to get started.

The change includes:
- Adding the new string resource and its Chinese translation.
- Updating the `MedicationPlanScreen` to display the instructional text with appropriate styling and spacing when the list of medication plans is empty.

Signed-off-by: 339 <339@339.im>